Add totalStorage field to School model (#49)

+++ src/file-on-student-assignment/file-on-student-assignment.service.ts
@@ -1,14 +1,22 @@
+import { TeacherOnSubjectRepository } from './../teacher-on-subject/teacher-on-subject.repository';
+import { SchoolRepository } from './../school/school.repository';
 import { AssignmentRepository } from './../assignment/assignment.repository';
 import { DeleteFileOnStudentAssignmentDto } from './dto/delete-file.dto';
 import { StudentOnAssignmentRepository } from './../student-on-assignment/student-on-assignment.repository';
 import { FileOnStudentAssignmentRepository } from './file-on-student-assignment.repository';
 import { GoogleStorageService } from './../google-storage/google-storage.service';
-import { Injectable, Logger, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  Logger,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import {
   CreateFileOnStudentAssignmentDto,
   GetFileOnStudentAssignmentByStudentOnAssignmentIdDto,
 } from './dto';
+import { Student, User } from '@prisma/client';
 
 @Injectable()
 export class FileOnStudentAssignmentService {
@@ -16,6 +24,9 @@ export class FileOnStudentAssignmentService {
   assignmentRepository: AssignmentRepository = new AssignmentRepository(
     this.prisma,
   );
+  teacherOnSubjectRepository: TeacherOnSubjectRepository =
+    new TeacherOnSubjectRepository(this.prisma);
+  schoolRepository: SchoolRepository = new SchoolRepository(this.prisma);
   fileOnStudentAssignmentRepository: FileOnStudentAssignmentRepository =
     new FileOnStudentAssignmentRepository(
       this.prisma,
@@ -28,10 +39,19 @@ export class FileOnStudentAssignmentService {
     private googleStorageService: GoogleStorageService,
   ) {}
 
-  async getFileByStudentOnAssignmentId(
+  async getFileByStudentOnAssignmentIdFromStudent(
     dto: GetFileOnStudentAssignmentByStudentOnAssignmentIdDto,
+    student: Student,
   ) {
     try {
+      const studentOnAssignment =
+        await this.studentOnAssignmentRepository.getById({
+          studentOnAssignmentId: dto.studentOnAssignmentId,
+        });
+
+      if (studentOnAssignment.studentId !== student.id) {
+        throw new ForbiddenException("You don't have permission to access");
+      }
       return await this.fileOnStudentAssignmentRepository.getByStudentOnAssignmentId(
         dto,
       );
@@ -41,27 +61,86 @@ export class FileOnStudentAssignmentService {
     }
   }
 
-  async createFileOnStudentAssignment(dto: CreateFileOnStudentAssignmentDto) {
+  async getFileByStudentOnAssignmentIdFromTeacher(
+    dto: GetFileOnStudentAssignmentByStudentOnAssignmentIdDto,
+    user: User,
+  ) {
     try {
-      const studnetOnAssignment =
+      const studentOnAssignment =
         await this.studentOnAssignmentRepository.getById({
           studentOnAssignmentId: dto.studentOnAssignmentId,
         });
 
-      return await this.fileOnStudentAssignmentRepository.create({
-        ...dto,
-        schoolId: studnetOnAssignment.schoolId,
-        subjectId: studnetOnAssignment.subjectId,
-        assignmentId: studnetOnAssignment.assignmentId,
-        studentId: studnetOnAssignment.studentId,
-      });
+      if (!studentOnAssignment) {
+        throw new NotFoundException('Student on assignment not found');
+      }
+
+      const teacherOnSubject =
+        await this.teacherOnSubjectRepository.getByTeacherIdAndSubjectId({
+          teacherId: user.id,
+          subjectId: studentOnAssignment.subjectId,
+        });
+
+      if (!teacherOnSubject) {
+        throw new ForbiddenException("You don't have permission to access");
+      }
+      return await this.fileOnStudentAssignmentRepository.getByStudentOnAssignmentId(
+        dto,
+      );
     } catch (error) {
       this.logger.error(error);
       throw error;
     }
   }
 
-  async deleteFileOnStudentAssignment(dto: DeleteFileOnStudentAssignmentDto) {
+  async createFileOnStudentAssignmentFromStudent(
+    dto: CreateFileOnStudentAssignmentDto,
+    student: Student,
+  ) {
+    try {
+      const studnetOnAssignment =
+        await this.studentOnAssignmentRepository.getById({
+          studentOnAssignmentId: dto.studentOnAssignmentId,
+        });
+
+      if (!studnetOnAssignment) {
+        throw new NotFoundException('Student on assignment not found');
+      }
+
+      if (studnetOnAssignment.studentId !== student.id) {
+        throw new ForbiddenException("You don't have permission to access");
+      }
+
+      const create = await this.fileOnStudentAssignmentRepository.create({
+        ...dto,
+        schoolId: studnetOnAssignment.schoolId,
+        subjectId: studnetOnAssignment.subjectId,
+        assignmentId: studnetOnAssignment.assignmentId,
+        studentId: studnetOnAssignment.studentId,
+      });
+
+      const school = await this.schoolRepository.getById({
+        schoolId: studnetOnAssignment.schoolId,
+      });
+
+      await this.schoolRepository.update({
+        query: { schoolId: studnetOnAssignment.schoolId },
+        body: {
+          totalStorage: school.totalStorage + create.size,
+        },
+      });
+
+      return create;
+    } catch (error) {
+      this.logger.error(error);
+      throw error;
+    }
+  }
+
+  async deleteFileOnStudentAssignmentFromStudnet(
+    dto: DeleteFileOnStudentAssignmentDto,
+    student: Student,
+  ) {
     try {
       const fileOnStudentAssignment =
         await this.fileOnStudentAssignmentRepository.getById({
@@ -70,6 +149,10 @@ export class FileOnStudentAssignmentService {
 
       if (!fileOnStudentAssignment) {
         throw new NotFoundException('File not found');
+      }
+
+      if (fileOnStudentAssignment.studentId !== student.id) {
+        throw new ForbiddenException("You don't have permission to access");
       }
       const assignment = await this.assignmentRepository.getAssignmentById({
         assignmentId: fileOnStudentAssignment.studentOnAssignmentId,
@@ -84,9 +167,68 @@ export class FileOnStudentAssignmentService {
           'This assignment is not allow to delete work',
         );
       }
-      return await this.fileOnStudentAssignmentRepository.delete({
+      const deleteFile = await this.fileOnStudentAssignmentRepository.delete({
         fileOnStudentAssignmentId: dto.fileOnStudentAssignmentId,
       });
+
+      const school = await this.schoolRepository.getById({
+        schoolId: fileOnStudentAssignment.schoolId,
+      });
+
+      await this.schoolRepository.update({
+        query: { schoolId: fileOnStudentAssignment.schoolId },
+        body: {
+          totalStorage: school.totalStorage - fileOnStudentAssignment.size,
+        },
+      });
+
+      return deleteFile;
+    } catch (error) {
+      this.logger.error(error);
+      throw error;
+    }
+  }
+
+  async deleteFileOnStudentAssignmentFromTeacher(
+    dto: DeleteFileOnStudentAssignmentDto,
+    user: User,
+  ) {
+    try {
+      const fileOnStudentAssignment =
+        await this.fileOnStudentAssignmentRepository.getById({
+          fileOnStudentAssignmentId: dto.fileOnStudentAssignmentId,
+        });
+
+      if (!fileOnStudentAssignment) {
+        throw new NotFoundException('File not found');
+      }
+
+      const teacherOnSubject =
+        await this.teacherOnSubjectRepository.getByTeacherIdAndSubjectId({
+          teacherId: user.id,
+          subjectId: fileOnStudentAssignment.subjectId,
+        });
+
+      if (!teacherOnSubject) {
+        throw new ForbiddenException("You don't have permission to access");
+      }
+
+      const deleteFile = await this.fileOnStudentAssignmentRepository.delete({
+        fileOnStudentAssignmentId: dto.fileOnStudentAssignmentId,
+      });
+
+      const school = await this.schoolRepository.getById({
+        schoolId: fileOnStudentAssignment.schoolId,
+      });
+
+      await this.schoolRepository.update({
+        query: { schoolId: fileOnStudentAssignment.schoolId },
+        body: {
+          totalStorage: school.totalStorage - fileOnStudentAssignment.size,
+        },
+      });
+
+      return deleteFile;
     } catch (error) {
       this.logger.error(error);
       throw error;
